requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
attrs==25.4.0
billiard==4.2.4
celery==5.3.4
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
cryptography==46.0.5
cssselect2==0.9.0
diff-match-patch==20241021
dj-database-url==2.1.0
Django==5.2.11
django-cors-headers==4.3.1
django-crispy-forms==2.5
django-extensions==4.1
django-guardian==3.2.0
django-import-export==4.4.0
django-unfold==0.78.1
et_xmlfile==2.0.0
freetype-py==2.3.0
frozenlist==1.8.0
gunicorn==21.2.0
html5lib==1.1
idna==3.11
kombu==5.6.2
lxml==6.0.2
multidict==6.7.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.9
pycairo==1.29.0
pycparser==3.0
pyHanko==0.33.0
pyhanko-certvalidator==0.29.1
PyJWT==2.11.0
pypdf==6.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.3
redis==5.0.1
reportlab==4.0.9
requests==2.31.0
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.5.1
tablib==3.9.0
tinycss2==1.5.1
twilio==8.10.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.12
yarl==1.22.0
